do not remove autogenerated.content for git

+++ imxTools/utils/helpers.py
@@ -4,7 +4,7 @@ from pathlib import Path
 def clear_directory(directory: Path):
     if directory.exists() and directory.is_dir():
         for item in directory.iterdir():
-            if item.is_file():
+            if item.is_file() and item.name != "generated.content":
                 item.unlink()  # Remove file
             elif item.is_dir():
                 clear_directory(item)  # Recursively clear subdirectory
